set up intent
I used Intent.putExtra() to add data such as name and date (which is a
string as of now).  Also added activities corresponding to the create
and group buttons, and their respective xml layout files and .java
screen files.  I implemented the onActivityResult method, which does
stuff for when the Create activity finishes.  I just can't figure out
why the hell the data of the new train doesn't show up correctly - it
adds a new train but it is blank